refactor l1 client
