+++ clients/tools/bridge/init.ts
@@ -1,9 +1,8 @@
-import { EthConfig } from "../config";
-import { Bridge, withBridgeClient } from "./bridge";
+import { EthConfig } from "../../config";
+import { L1Client, withL1Client } from "../../client";
 import { encodeL1address } from "web3subscriber/src/addresses";
 
 const Tokens = require("./tokenlist");
-const TokenInfo = require("../../build/contracts/Token.json");
 const Secrets = require("../../.secrets");
 const fs = require("fs");
 const path = require("path");
@@ -19,21 +18,13 @@ function crunchTokens() {
 async function main(config_name: string) {
   console.log("start calling");
   let config = EthConfig[config_name](Secrets);
-  let address = TokenInfo.networks[config.device_id].address;
   try {
-    withBridgeClient(config, false, async (bridge: Bridge) => {
-      /*
-      let token = bridge.web3.getContract(
-        TokenInfo,
-        address,
-        bridge.web3.getDefaultAccount()
-      );
-      */
-
+    withL1Client(config, false, async (l1client: L1Client) => {
+      let bridge = l1client.getBridgeContract();
       let output: any = {};
       let index = 4;
 
-      console.log("Testing bridge [id=%s]", bridge.getChainIdHex());
+      console.log("Testing bridge [id=%s]", l1client.getChainIdHex());
       for (let tokenUid of crunchTokens()) {
         console.log("Adding token uid: 0x", tokenUid.toString(16));
         let tx = await bridge.addToken(tokenUid);
