+++ clients/tools/bridge/status.js
