--- clients/types/bridge.ts
@@ -1,12 +1,0 @@
-export interface BridgeInfo {
-    chain_id: string;
-    amount_token: string;
-    amount_pool: string;
-    owner: string;
-    merkle_root: string;
-    rid: string;
-  }
-
-  export interface TokenInfo {
-    token_uid: string;
-  }
